requirements.txt
python-dotenv==1.1.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
beautifulsoup4==4.13.4